Remove ios code
I cannot get the ios code to properly include the headers and link to
the frameworks. At this point, I give up. I don't want to block releases
of GitJournal in ios any longer so for now I'm keeping the ios code in
git_bindings in GitJournal and just changing the platform channel name.

It's an ugly hack, but I've invested way too much time into this
already. It's time to ask an ios developer for help.